fixed clippy and large files
